// client/swagger/model/Bundle.ts
/**
 * LearnIt
 * Temporal Api for LearnIt.
 *
 * OpenAPI spec version: 0.1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Bundle {
    id: string;

    languageId: string;

    name: string;

    customBundle?: boolean;

    favorite?: boolean;

    csvStringUrl?: string;

}