// client/swagger/model/Language.ts
/**
 * LearnIt
 * Temporal Api for LearnIt.
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Language {
    id: string;

    name1: string;

    shortName1: string;

    image1?: string;

    name2: string;

    shortName2: string;

    image2?: string;

}